% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunburst.R
\name{renderSunburst}
\alias{renderSunburst}
\title{Widget render function for use in Shiny}
\usage{
renderSunburst(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

