% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunburst.R
\name{sunburstOutput}
\alias{sunburstOutput}
\title{Widget output function for use in Shiny}
\usage{
sunburstOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

