% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2b.R
\name{sund2b}
\alias{sund2b}
\title{Sunburst Using 'd2b'}
\usage{
sund2b(data = NULL, valueField = "size", width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{data}{data in csv source,target form or in
nested d3 JSON hierarchy with `{name:...,  children:[];}`.  \code{list}, \code{character},
or \code{connection} data will be assumed to be \code{JSON}.
\code{data.frame} data will be assumed to be \code{csvdata} and converted
to \code{JSON} by \code{sunburstR:::csv_to_hier()}.}
}
\description{
Create interactive sunburst chart with the 'd2b' charting library.
}
