% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{add_shiny}
\alias{add_shiny}
\title{Add Shiny Events}
\usage{
add_shiny(sunburst = NULL)
}
\arguments{
\item{sunburst}{\code{sunburst} htmlwidget to which you would
like to add event handling}
}
\value{
\code{sunburst} htmlwidget
}
\description{
Add Shiny Events
}

