% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sunburst.R
\name{sunburst}
\alias{sunburst}
\title{htmlwidget for d3.js sequence sunburst diagrams}
\usage{
sunburst(csvdata = NULL, jsondata = NULL, width = NULL, height = NULL)
}
\description{
\href{https://gist.github.com/kerryrodden/7090426}{Sequences sunburst} diagrams provide
   an interactive method of exploring sequence data, such as website navigation paths.
}

